Require use of unwritable input types in cooperative vectors
For matrix and vector ref inputs, the use of the *Impl template allowed passing in writable RWMatrixRef and RWVectorRef. This wasn't desired and didn't actually work with the implementation.

This requires read-only type parameters to the intrinsics with the exception of the output matrix in OuterProductAccumulate. This has the consequence of removing RWVectorRef entirely and changing the conceptual API definition to remove unneeded template parameters and change how the Refs are defined for implementation reasons, but also to simplify the conceptual picture of the implementaiton.